plot_gap in separate file, update bilinear form
