+++ python/dolfinx_contact/helpers.py
@@ -9,7 +9,6 @@ import numpy
 import ufl
 from contextlib import ExitStack
 from petsc4py import PETSc
-from matplotlib import pyplot as plt
 
 __all__ = ["lame_parameters", "epsilon", "sigma_func", "convert_mesh", "plot_gap"]
 
@@ -168,37 +167,3 @@ def rigid_motions_nullspace(V):
     _x = [basis[i] for i in range(dim)]
     nsp = PETSc.NullSpace().create(vectors=_x)
     return nsp
-
-
-# Visualise the gap
-def plot_gap(mesh, contact, tag, gap):
-    gdim = mesh.geometry.dim
-    fdim = mesh.topology.dim - 1
-    mesh_geometry = mesh.geometry.x
-    if tag == 0:
-        facets = contact.facet_0()
-        facets_opp = contact.facet_1()
-    else:
-        facets = contact.facet_1()
-        facets_opp = contact.facet_0()
-
-    # Draw facets on opposite surface
-    plt.figure()
-    for facet in facets_opp:
-        facet_geometry = dolfinx.cpp.mesh.entities_to_geometry(mesh, fdim, [facet], False)
-        coords = mesh_geometry[facet_geometry][0]
-        plt.plot(coords[:, 0], coords[:, 1], color="black")
-    num_facets = len(facets)
-    for i in range(num_facets):
-        facet = facets[i]
-        facet_geometry = dolfinx.cpp.mesh.entities_to_geometry(mesh, fdim, [facet], False)
-        coords = mesh_geometry[facet_geometry][0]
-        plt.plot(coords[:, 0], coords[:, 1], color="black")
-        qp = contact.qp_phys(tag, i)
-        num_qp = qp.shape[0]
-        for q in range(num_qp):
-            g = gap[i, q * gdim:(q + 1) * gdim]
-            x = [qp[q, 0], qp[q, 0] + g[0]]
-            y = [qp[q, 1], qp[q, 1] + g[1]]
-            plt.plot(x, y)
-    plt.savefig(f"gap_{tag}.svg")
